requirements.txt
Flask==3.1.2
SQLAlchemy==2.0.44
PyYAML==6.0.3
python-vlc==3.0.21203
mutagen==1.47.0
gunicorn==21.2.0; platform_system != "Windows"
RPi.GPIO==0.7.1; platform_system == "Linux"
ttkbootstrap==1.10.1
requests==2.32.3
